Move input transform
