+++ docs/source/api/flash.rst
@@ -12,6 +12,6 @@ flash
     ~flash.core.data.callback.FlashCallback
     ~flash.core.data.io.output_transform.OutputTransform
     ~flash.core.data.io.output.Output
-    ~flash.core.data.input_transform.InputTransform
+    ~flash.core.data.io.input_transform.InputTransform
     ~flash.core.model.Task
     ~flash.core.trainer.Trainer
