+++ flash/audio/classification/data.py
@@ -29,8 +29,8 @@ from flash.audio.classification.input_transform import AudioClassificationInputT
 from flash.core.data.callback import BaseDataFetcher
 from flash.core.data.data_module import DataModule
 from flash.core.data.data_pipeline import DataPipelineState
-from flash.core.data.input_transform import INPUT_TRANSFORM_TYPE
 from flash.core.data.io.input import Input
+from flash.core.data.io.input_transform import INPUT_TRANSFORM_TYPE
 from flash.core.data.utilities.paths import PATH_TYPE
 from flash.core.registry import FlashRegistry
 from flash.core.utilities.stages import RunningStage
