+++ flash/core/model.py
@@ -38,8 +38,8 @@ from torch.utils.data import DataLoader, Sampler
 
 import flash
 from flash.core.data.data_pipeline import DataPipeline, DataPipelineState
-from flash.core.data.input_transform import InputTransform
 from flash.core.data.io.input import InputBase, ServeInput
+from flash.core.data.io.input_transform import InputTransform
 from flash.core.data.io.output import Output
 from flash.core.data.io.output_transform import OutputTransform
 from flash.core.data.process import Deserializer, DeserializerMapping
