+++ flash/core/utilities/types.py
@@ -32,7 +32,7 @@ METRICS_TYPE = TypeVar("METRICS_TYPE", Metric, Mapping, Sequence, None)
 DESERIALIZER_TYPE = TypeVar("DESERIALIZER_TYPE", "flash.core.data.process.Deserializer", None)
 INPUT_TRANSFORM_TYPE = TypeVar(
     "INPUT_TRANSFORM_TYPE",
-    Type["flash.core.data.input_transform.InputTransform"],
+    Type["flash.core.data.io.input_transform.InputTransform"],
     Callable,
     Tuple[Union[LightningEnum, str], Dict[str, Any]],
     Union[LightningEnum, str],
