+++ flash/graph/classification/input_transform.py
@@ -16,8 +16,8 @@ from typing import Any, Callable, Dict, List
 
 from torch.utils.data.dataloader import default_collate
 
-from flash.core.data.input_transform import InputTransform
 from flash.core.data.io.input import DataKeys
+from flash.core.data.io.input_transform import InputTransform
 from flash.core.utilities.imports import _GRAPH_AVAILABLE
 
 if _GRAPH_AVAILABLE:
