+++ flash/pointcloud/detection/data.py
@@ -17,8 +17,8 @@ from torch.utils.data import Dataset
 
 from flash.core.data.data_module import DataModule
 from flash.core.data.data_pipeline import DataPipelineState
-from flash.core.data.input_transform import InputTransform
 from flash.core.data.io.input import BaseDataFormat, Input
+from flash.core.data.io.input_transform import InputTransform
 from flash.core.utilities.stages import RunningStage
 from flash.core.utilities.types import INPUT_TRANSFORM_TYPE
 from flash.pointcloud.detection.input import PointCloudObjectDetectorDatasetInput
