+++ flash/tabular/regression/model.py
@@ -17,8 +17,8 @@ from typing import Any, Callable, Dict, List, Optional, Tuple, Type
 import torch
 from torch.nn import functional as F
 
-from flash.core.data.input_transform import InputTransform
 from flash.core.data.io.input import DataKeys, ServeInput
+from flash.core.data.io.input_transform import InputTransform
 from flash.core.regression import RegressionTask
 from flash.core.serve import Composition
 from flash.core.utilities.imports import _TABULAR_AVAILABLE, requires
