+++ tests/core/data/test_callbacks.py
@@ -18,8 +18,8 @@ from torch import tensor
 
 from flash.core.data.callback import BaseDataFetcher
 from flash.core.data.data_module import DataModule
-from flash.core.data.input_transform import InputTransform
 from flash.core.data.io.input import Input
+from flash.core.data.io.input_transform import InputTransform
 from flash.core.utilities.stages import RunningStage
